[BOOT] Add an ISO remastering script.
This is useful in the context of e.g. 3rd-party testers who want to
quickly add/remove/modify any file in a ReactOS ISO image and remastering
it without the need of our whole build environment.

TODO: Test/exercise this script and improve it, then add an equivalent
BASH script.