TEMP: HERM SYNCH FIXES
[WINESYNC] Minor formatting.

[WINESYNC] Fix some apparent problems with commits manipulation with pygit2.

[WINESYNC] Explicitly use posix paths for git actions.

Whenever using git, be it on Windows CMD, git-bash on Windows, or in
*nix systems, git appears to always use posix-like paths internally.
So, enforce them with posixpath when building/concatenating paths that
are going to be passed to pygit2 functions.

Otherwise, exceptions like the following one happens:
```
  File "./winesync.py", line 296, in revert_staged_patchset
    self.reactos_index.remove(os.path.join(self.staged_patch_dir, patch_file_name))
[... skipped ...]
OSError: index does not contain sdk\tools\winesync\setupapi_staging\0002-wine-staging-4.0-setupapi_winetest.patch at stage 0
```
(The git index actually contains the specified path, but in "posix" form
with slash-separators.)

On *nix platforms, these changes should not matter.

[WINESYNC] Use --no-autoconf otherwise running time is too long.

[WINESYNC] Allow using the "new" Wine-Staging patchinstall.py script.

Wine-Staging switched to staging/patchinstall.py , removing the
deprecated patches/patchinstall.sh in Feb.16, 2023 commit
https://github.com/wine-staging/wine-staging/commit/c1b4af92f74d7bd330003d39d0bb1f966fdb70a9
just before the v8.2 release.

In order to maintain interoperability between older and newer
Wine-Staging versions, try to run first the new script; if it fails,
fall back to the older script.

[WINESYNC] Elegantly handle empty 'files' and 'directories' lists.

Expand on Timo's commit 4b5a55516.
We may encounter cases where either the 'directories' or 'files' lists
in the .cfg YAML files are empty, and we don't want the script to throw
an exception in that case.

Furthermore, explicitly check for such empty lists when calling PyGit2
index.add_all(...) function, because if it's called on a None or empty
list, _all_ untracked files in the selected git repository get added,
which is not what we want there.

[WINESYNC] Make the wine-staging tag optional, in which case Wine-Staging is not used.

[WINESYNC] Use the same 'winesync' author as for all the other commits the script generates.

This now correctly sets the commit author as 'winesync' (before it was
the committer's developer name) for the following types of commits:

`[WINESYNC]: revert wine-staging patchset for <MODULE_NAME>`

and

`[WINESYNC]: <MODULE_NAME> is now in sync with wine-staging <WINE_TAG>`

[WINESYNC] When initially reverting wine-staging patches, ignore whitespace.

This fixes in particular problems when reverting created new files (i.e.
they are deleted). I suspect this is due to an endline-format "expectation"
from the git apply command.